[froude_krylov] Fixed cast error detected by Jenkins

+++ code/model_wrappers/src/FroudeKrylovForceModel.cpp
@@ -22,7 +22,7 @@ SurfaceForceModel::DF FroudeKrylovForceModel::dF(const FacetIterator& that_facet
     {
         eta += body.intersector->all_immersions.at(*it);
     }
-    if (not(that_facet->vertex_index.empty())) eta /= that_facet->vertex_index.size();
+    if (not(that_facet->vertex_index.empty())) eta /= (double)that_facet->vertex_index.size();
 
     const double pdyn = env.w->get_dynamic_pressure(env.rho,env.g,C,env.k,eta,t);
     return DF(-pdyn*dS,C.v);
